collage/requirements.txt
boto3==1.17.17
botocore==1.20.17
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
pathlib==1.0.1
Pillow==8.1.0
python-dateutil==2.8.1
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
